The wrapped column counts the characters that fit, not the ones a rounded advance suggests
viewportCols() divided the message column's width by
QFontMetrics::averageCharWidth(), and every integer advance Qt offers is a
rounded one: at the reference fixed-pitch face both averageCharWidth() and
horizontalAdvance() answer 7 at the shipped 9 pt where the advance is 7.21875,
below the true advance at 15 of the 27 point sizes the zoom offers — the same 15
QFontMetrics::height() was short at, which is why this is the horizontal sibling
of the line-pitch fix and survived it untouched.

Always On gives a record ceil(chars / cols) rows and then clips its text to them,
so an overcounted cols costs the record a line. Measured in a 880 px viewport at
9 pt: a 379 px column counted 54 characters where 52 fit, a 53-character record
was given one row, and the two characters past the fold were drawn nowhere — no
ellipsis and no tooltip, because a wrapped message deliberately offers neither.
Over 34,587 size-and-width pairs the old expression overcounted at 15,734 of
them, worst case by 22 columns.

The advance is now fractional and the division floored. Rounding the advance up
instead is the mirror fault and was not taken: qCeil buys 47 columns where 52
fit, which is a blank line hung off every wrapped record. The kMinWrapCols floor
had to move in the same change — minWrapWidth() multiplied the same truncated
advance out, and viewportCols() re-raises the count to it with a qMax, so at the
floor width the clamp still overcounted by one at those same 15 sizes with the
division alone fixed.

The three new cases in tst_logview are the shape the line-pitch case established:
sweep the point sizes for one where the metrics actually diverge and skip if none
does, then read the answer off a rendered viewport, since none of this is
observable any other way — every widget holds the right value throughout and only
the rows the record was drawn into are wrong. They measure the fit through a real
QTextLayout in digits, sweep every size at four widths, and read the three record
lengths that straddle the fold off the pixels. All three fail at the previous
commit.

Two findings from the work are recorded in bugs.md as 21 and 22 rather than
fixed, because the answer to both is a design decision. 21: the height model
assumes the primary face's advance, so a record of CJK or emoji resolves through
a wider fallback face and loses lines the same silent way — 40 x U+4E2D at 9 pt
is allotted one row and laid out in two. 22: QTextLine's break test allows for
the last glyph's right bearing, so a line ending on an overhanging glyph ('W' at
most sizes, 'x' at 7 pt) still breaks one character early, which no
width-over-advance arithmetic can see. Both want the same answer — measure such
records rather than count them — and that puts a shaping pass back on the path
7.1.1 exists to keep free of one.

+++ src/ui/LogView.h
@@ -402,8 +402,12 @@ private:
     // AlwaysOn support: characters that fit across the message column, block
     // measurement from decoded text, and keeping the visible blocks measured.
     int viewportCols() const;
-    // One character's advance, and the pixel floor kMinWrapCols of them come to.
-    int charAdvance() const;
+    // One character's advance, and the pixel floor kMinWrapCols of them come to. The
+    // advance is FRACTIONAL and must stay so: every integer advance Qt offers —
+    // QFontMetrics::averageCharWidth() and QFontMetrics::horizontalAdvance() alike —
+    // rounds, and a rounded-down advance overcounts the columns that fit, which costs a
+    // wrapped record a row and clips its tail (ARCHITECTURE.md §7.1.1).
+    qreal charAdvance() const;
     int minWrapWidth() const;
     void ensureEstimatorBound();
     void measureBlock(int block);
